# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/me": {
      "get": {
        "description": "Returns profile information of the currently logged-in user.",
        "operationId": "UserController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the authenticated user profile.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "User not authenticated."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get authenticated user profile",
        "tags": [
          "User"
        ]
      },
      "delete": {
        "description": "Deletes the logged-in user account.",
        "operationId": "UserController_deleteAccount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Account deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAccountResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "User not authenticated."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Delete current user account",
        "tags": [
          "User"
        ]
      }
    },
    "/users": {
      "get": {
        "description": "Returns a list of all registered users. Admin access required.",
        "operationId": "UserController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the list of all users.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllUsersResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Only admin can access this endpoint."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get all users (Admin)",
        "tags": [
          "User"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Returns the profile of a specific user using their ID. Admin only.",
        "operationId": "UserController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns user details for the given ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Only admin can access this endpoint."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get user by ID (Admin)",
        "tags": [
          "User"
        ]
      },
      "delete": {
        "description": "Deletes a specific user account. Only Admin can perform this action.",
        "operationId": "UserController_deleteById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User account deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAccountResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Only admin can access this endpoint."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Delete user by ID (Admin)",
        "tags": [
          "User"
        ]
      }
    },
    "/users/{id}/role": {
      "patch": {
        "description": "Allows admin to assign or update the role (ADMIN/USER) of a user.",
        "operationId": "UserController_assignRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User role updated successfully."
          },
          "400": {
            "description": "Invalid role value."
          },
          "403": {
            "description": "Only admin can access this endpoint."
          },
          "404": {
            "description": "User not found."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Update user role (Admin)",
        "tags": [
          "User"
        ]
      }
    },
...
  },
  "info": {
    "title": "Task Management System API",
    "description": "A fully functional Task Management System backend built with **NestJS**,\n      providing secure authentication, role-based access control, and task and user management endpoints.\n      ",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "accessToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserTaskDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "object",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "PAUSED",
              "COMPLETED",
              "CANCELED"
            ],
            "example": "NOT_STARTED"
          },
          "assignedToUserId": {
            "type": "object",
            "nullable": true
          },
          "assignedAt": {
            "type": "object",
            "nullable": true
          },
          "completedAt": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "status",
          "assignedToUserId",
          "assignedAt",
          "completedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user ID",
            "example": "c3f0f1f2-2d3a-4c1e-a9e6-1b1234567890"
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "john@example.com"
          },
          "name": {
            "type": "string",
            "description": "Full name of the user",
            "example": "John Doe"
          },
          "role": {
            "type": "string",
            "description": "Role of the user",
            "example": "USER",
            "enum": [
              "ADMIN",
              "USER"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Account creation timestamp",
            "example": "2025-01-25T10:15:30.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last updated timestamp",
            "example": "2025-01-26T09:20:15.000Z"
          },
          "tasks": {
            "description": "Tasks assigned to the user",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserTaskDto"
            }
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "createdAt",
          "updatedAt",
          "tasks"
        ]
      },
      "DeleteAccountResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "UserSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "c3f0f1f2-2d3a-4c1e-a9e6-1b1234567890"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER"
            ],
            "example": "USER"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-25T10:15:30.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-26T09:20:15.000Z"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
      "AllUsersResponseDto": {
        "type": "object",
        "properties": {
          "users": {
            "description": "List of all registered users",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserSummaryDto"
            }
          }
        },
        "required": [
          "users"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "description": "Role to assign to the user",
            "example": "ADMIN",
            "enum": [
              "ADMIN",
              "USER"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
...
    }
  }
}